fix(backend): rss RETURNING 列加 s. 别名 + scheduled_tasks $4 ::bigint 消歧 — 共享 PG 报错治理 (2026-09-14 PG 审计)
- rss/store.go: ClaimDueSources 的 UPDATE ... FROM claimed 里 RETURNING
  裸列名与 FROM 表冲突, PG 报 'column reference "id" is ambiguous'
  (本地 llm-gateway-pg pocket 库每分钟数条)。新增 sourceColsReturning
  (s. 前缀) 仅供该语句使用, 普通 SELECT 的 sourceCols 不动。
- scheduledtask/store.go: UpdateTaskAfterRun 的 $4 同时用作
  next_run_at(bigint) 与 CASE WHEN $4 = 0(整数字面量), pgx 零 OID
  推断报 'inconsistent types deduced for parameter $4: integer versus
  bigint'。哨兵比较改 $4::bigint = 0, 显式钉住参数类型。
- 两句均经 PREPARE 零类型推断路径正/阴对照验证; go build +
  rss/scheduledtask 测试全绿; 已随 opencode-pocket:pocket-opp 重建部署。
- 不改表结构, 不动其他仓库。

+++ backend/internal/scheduledtask/store.go
@@ -396,7 +396,7 @@ UPDATE scheduled_tasks SET
     run_count = run_count + 1,
     next_run_at = $4,
     lease_until = 0,
-    enabled = CASE WHEN $4 = 0 THEN FALSE ELSE enabled END,
+    enabled = CASE WHEN $4::bigint = 0 THEN FALSE ELSE enabled END,
     updated_at = $1
  WHERE id = $5
 `, unixNow(), string(status), errMsg, nextRunAt, taskID)
